+++ test/mix/tasks/ecto/pow_assent.ecto.gen.migration_test.exs
@@ -3,32 +3,25 @@ defmodule Mix.Tasks.PowAssent.Ecto.Gen.MigrationTest do
 
   alias Mix.Tasks.PowAssent.Ecto.Gen.Migration
 
-  defmodule Repo do
-    def __adapter__, do: true
-    def config, do: [priv: "tmp/#{inspect(Migration)}", otp_app: :pow_assent]
+  @migrations_path "migrations"
+
+  setup context do
+    {:ok, options: ["-r", inspect(context.repo)]}
   end
 
-  @tmp_path        Path.join(["tmp", inspect(Migration)])
-  @migrations_path Path.join([@tmp_path, "migrations"])
-  @options         ["-r", inspect(Repo)]
+  test "generates migration", context do
+    File.cd!(context.tmp_path, fn ->
+      Migration.run(context.options)
 
-  setup do
-    File.rm_rf!(@tmp_path)
-    File.mkdir_p!(@tmp_path)
+      assert_received {:mix_shell, :info, ["* creating ./migrations"]}
+      assert_received {:mix_shell, :info, ["* creating ./migrations/" <> _]}
 
-    :ok
-  end
-
-  test "generates migration" do
-    File.cd!(@tmp_path, fn ->
-      Migration.run(@options)
-
-      assert [migration_file] = File.ls!(@migrations_path)
+      assert [_, migration_file] = @migrations_path |> File.ls!() |> Enum.sort()
       assert String.match?(migration_file, ~r/^\d{14}_create_user_identities\.exs$/)
 
       file = @migrations_path |> Path.join(migration_file) |> File.read!()
 
-      assert file =~ "defmodule #{inspect(Repo)}.Migrations.CreateUserIdentities do"
+      assert file =~ "defmodule #{inspect(context.repo)}.Migrations.CreateUserIdentities do"
       assert file =~ "create table(:user_identities)"
       assert file =~ "add :provider, :string, null: false"
       assert file =~ "add :uid, :string, null: false"
@@ -37,23 +30,24 @@ defmodule Mix.Tasks.PowAssent.Ecto.Gen.MigrationTest do
     end)
   end
 
-  test "doesn't make duplicate migrations" do
-    File.cd!(@tmp_path, fn ->
-      Migration.run(@options)
+  test "doesn't make duplicate migrations", context do
+    File.cd!(context.tmp_path, fn ->
+      Migration.run(context.options)
 
       assert_raise Mix.Error, "migration can't be created, there is already a migration file with name CreateUserIdentities.", fn ->
-        Migration.run(@options)
+        Migration.run(context.options)
       end
     end)
   end
 
-  test "generates with binary_id" do
-    options = @options ++ ~w(--binary-id)
+  @tag repo: Repo
+  test "generates with binary_id", context do
+    options = context.options ++ ~w(--binary-id)
 
-    File.cd!(@tmp_path, fn ->
+    File.cd!(context.tmp_path, fn ->
       Migration.run(options)
 
-      assert [migration_file] = File.ls!(@migrations_path)
+      assert [_, migration_file] = @migrations_path |> File.ls!() |> Enum.sort()
       assert String.match?(migration_file, ~r/^\d{14}_create_user_identities\.exs$/)
 
       file = @migrations_path |> Path.join(migration_file) |> File.read!()